First Draft
This image shows the different restaurants located in Bowling Green, KY. 317 different restaurants are on this map. The final map will show how many restaurants are privately owned (local) or are corporately owned (part of a franchise). I may use a heatmap to demonstrate this and also use different shapes to represent whether the restaurant is brick/mortar (stationed at a single physical location) or is a food truck.